refactor rust_core module to separate files
